+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,7 +7,8 @@ development
 - middleware: number normalization: AT362828E should be returned as AT362828T for querying at OPS
 - middleware/ui: ops upstream datasource crawler for fetching complete list of publication numbers, user interface
 - ui: fix “full-cycle” for firefox
-- middleware/ui: numberlist crawler for depatisnet
+- middleware/ui: numberlist crawler for DEPATISnet
+- middleware/ui: numberlist crawler for FulltextPRO
 
 0.84.0
 ======